+++ demo/test/components/previews/primer/docs/auto_complete_item_preview.rb
@@ -1,6 +1,6 @@
 module Primer
   module Docs
-    class ItemPreview < ViewComponent::Preview
+    class AutoCompleteItemPreview < ViewComponent::Preview
       def default; end
 
     end