+++ demo/test/components/previews/primer/docs/octicon_component_preview.rb
@@ -1,6 +1,8 @@
 module Primer
   module Docs
     class OcticonComponentPreview < ViewComponent::Preview
+      include Primer::ViewHelper
+      
       def default; end
 
       def medium; end