+++ test/application_system_test_case.rb
@@ -23,6 +23,7 @@ class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
     component_uri = self.class.name.gsub("Test", "").gsub("Integration", "").underscore
 
     visit("/rails/view_components/primer/#{component_uri}/#{preview_name}")
+
     assert_accessible(page)
   end